Update module structure of io and bit change on seed type

+++ src/kinmu_lib/seed.rs
@@ -3,10 +3,10 @@
 use rand::rngs::StdRng;
 use rand::{RngCore, SeedableRng};
 
-pub fn gen_rng_from_seed(seed: usize) -> Box<dyn RngCore> {
+pub fn gen_rng_from_seed(seed: u64) -> Box<dyn RngCore> {
     if seed == 0 {
         Box::new(rand::thread_rng())
     } else {
-        Box::new(StdRng::seed_from_u64(seed as u64))
+        Box::new(StdRng::seed_from_u64(seed))
     }
 }
